requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
